+++ Assets/ScriptableRenderLoop/HDRenderPipeline/PostProcess/PostProcessing.Settings.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1af1b2448dc9dd04fb678c5a50a8a515
+timeCreated: 1486116366
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
